test: speed up tests with compiled JAX
