+++ tests/test_CHO.py
@@ -17,16 +17,16 @@
 """Tests for C-H-O systems"""
 
 import logging
-from typing import Mapping
+from collections.abc import Mapping
 
 import numpy as np
+import pytest
 from jaxtyping import ArrayLike
 
-from atmodeller import __version__, debug_logger
+from atmodeller import debug_logger
 from atmodeller.classes import InteriorAtmosphere
-from atmodeller.containers import ConstantFugacityConstraint, Planet, Species, SpeciesCollection
+from atmodeller.containers import Planet, Species, SpeciesCollection
 from atmodeller.interfaces import FugacityConstraintProtocol, SolubilityProtocol
-from atmodeller.mytypes import NumpyArrayFloat
 from atmodeller.output import Output
 from atmodeller.solubility import get_solubility_models
 from atmodeller.thermodata import IronWustiteBuffer
@@ -35,275 +35,19 @@ from atmodeller.utilities import earth_oceans_to_hydrogen_mass
 logger: logging.Logger = debug_logger()
 logger.setLevel(logging.WARNING)
 
-RTOL: float = 1.0e-6
+RTOL: float = 1.0e-8
 """Relative tolerance"""
-ATOL: float = 1.0e-6
+ATOL: float = 1.0e-8
 """Absolute tolerance"""
+TOLERANCE: float = 5.0e-2
+"""Tolerance of log output to satisfy comparison with FactSage and FastChem"""
 
 solubility_models: Mapping[str, SolubilityProtocol] = get_solubility_models()
 
-
-def test_version():
-    """Test version."""
-    assert __version__ == "0.5.0"
-
-
-def test_H2O(helper) -> None:
-    """Tests a single species (H2O)."""
-
-    H2O_g: Species = Species.create_gas(
-        "H2O_g", solubility=solubility_models["H2O_peridotite_sossi23"]
-    )
-
-    species: SpeciesCollection = SpeciesCollection((H2O_g,))
-    planet: Planet = Planet()
-    interior_atmosphere: InteriorAtmosphere = InteriorAtmosphere(species)
-
-    oceans: ArrayLike = 2
-    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
-    mass_constraints: dict[str, ArrayLike] = {"H": h_kg}
-
-    interior_atmosphere.solve(
-        planet=planet,
-        mass_constraints=mass_constraints,
-    )
-    output: Output = interior_atmosphere.output
-    solution: dict[str, ArrayLike] = output.quick_look()
-
-    target: dict[str, float] = {
-        "H2O_g": 1.0312913336898137,
-    }
-
-    assert helper.isclose(solution, target, rtol=RTOL, atol=ATOL)
-
-
-def test_H_fO2(helper) -> None:
-    """Tests H2-H2O at the IW buffer with H2O solubility."""
-
-    H2O_g: Species = Species.create_gas(
-        "H2O_g", solubility=solubility_models["H2O_peridotite_sossi23"]
-    )
-    H2_g: Species = Species.create_gas("H2_g")
-    O2_g: Species = Species.create_gas("O2_g")
-
-    species: SpeciesCollection = SpeciesCollection((H2O_g, H2_g, O2_g))
-    planet: Planet = Planet()
-    interior_atmosphere: InteriorAtmosphere = InteriorAtmosphere(species)
-
-    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {"O2_g": IronWustiteBuffer()}
-
-    oceans: float = 1
-    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
-    mass_constraints: dict[str, ArrayLike] = {"H": h_kg}
-
-    interior_atmosphere.solve(
-        planet=planet,
-        fugacity_constraints=fugacity_constraints,
-        mass_constraints=mass_constraints,
-    )
-    output: Output = interior_atmosphere.output
-    solution: dict[str, ArrayLike] = output.quick_look()
-
-    target: dict[str, float] = {
-        "H2O_g": 0.2570800742364775,
-        "H2_g": 0.2491511264610601,
-        "O2_g": 8.838513516896038e-08,
-    }
-
-    assert helper.isclose(solution, target, rtol=RTOL, atol=ATOL)
-
-
-def test_H_fO2_fH2(helper) -> None:
-    """Tests H2-H2O at the IW buffer with H2O solubility and mixed fugacity constraints."""
-
-    H2O_g: Species = Species.create_gas(
-        "H2O_g", solubility=solubility_models["H2O_peridotite_sossi23"]
-    )
-    H2_g: Species = Species.create_gas("H2_g")
-    O2_g: Species = Species.create_gas("O2_g")
-
-    species: SpeciesCollection = SpeciesCollection((H2O_g, H2_g, O2_g))
-    planet: Planet = Planet()
-    interior_atmosphere: InteriorAtmosphere = InteriorAtmosphere(species)
-
-    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {
-        "O2_g": IronWustiteBuffer(np.array([-1, 0, 1], dtype=np.float64)),
-        "H2_g": ConstantFugacityConstraint(np.array([1.0e-8, 1.0e-7, 1.0e-6], dtype=np.float64)),
-    }
-
-    interior_atmosphere.solve(
-        planet=planet,
-        fugacity_constraints=fugacity_constraints,
-    )
-    output: Output = interior_atmosphere.output
-    solution: dict[str, ArrayLike] = output.quick_look()
-
-    target: dict[str, ArrayLike] = {
-        "H2O_g": np.array([3.262913506271090e-09, 1.031823848794260e-07, 3.262913506271089e-06]),
-        "H2_g": np.array([1.000000000000005e-08, 9.999999999999959e-08, 1.000000000000000e-06]),
-        "O2_g": np.array([8.838513516896060e-09, 8.838513516896038e-08, 8.838513516896018e-07]),
-    }
-
-    assert helper.isclose(solution, target, rtol=RTOL, atol=ATOL)
-
-
-def test_H_fO2_batch_temperature(helper) -> None:
-    """Tests H2-H2O at the IW buffer with H2O solubility for a range of surface temperatures."""
-
-    H2O_g: Species = Species.create_gas(
-        "H2O_g", solubility=solubility_models["H2O_peridotite_sossi23"]
-    )
-    H2_g: Species = Species.create_gas("H2_g")
-    O2_g: Species = Species.create_gas("O2_g")
-
-    species: SpeciesCollection = SpeciesCollection((H2O_g, H2_g, O2_g))
-    interior_atmosphere: InteriorAtmosphere = InteriorAtmosphere(species)
-
-    # Number of surface temperatures is different to number of species to test array shapes work.
-    surface_temperatures: NumpyArrayFloat = np.array([1500, 2000, 2500, 3000], dtype=np.float64)
-    planet: Planet = Planet(surface_temperature=surface_temperatures)
-
-    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {"O2_g": IronWustiteBuffer()}
-
-    oceans: float = 1
-    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
-    mass_constraints: dict[str, ArrayLike] = {"H": h_kg}
-
-    interior_atmosphere.solve(
-        planet=planet,
-        fugacity_constraints=fugacity_constraints,
-        mass_constraints=mass_constraints,
-    )
-    output: Output = interior_atmosphere.output
-    solution: dict[str, ArrayLike] = output.quick_look()
-
-    target: dict[str, ArrayLike] = {
-        "H2O_g": np.array(
-            [
-                2.566653037020448e-01,
-                2.570800742364757e-01,
-                2.572178041535549e-01,
-                2.572746043480848e-01,
-            ]
-        ),
-        "H2_g": np.array(
-            [
-                3.133632393608037e-01,
-                2.491511264610584e-01,
-                2.265704456625875e-01,
-                2.199521409043987e-01,
-            ]
-        ),
-        "O2_g": np.array(
-            [
-                2.394194493859141e-12,
-                8.838513516896038e-08,
-                4.544970468047975e-05,
-                2.739422634823809e-03,
-            ]
-        ),
-    }
-
-    assert helper.isclose(solution, target, rtol=RTOL, atol=ATOL)
-
-
-def test_H_fO2_batch_fO2_shift(helper) -> None:
-    """Tests H2-H2O at the IW buffer with H2O solubility for a range of fO2 shifts."""
-
-    H2O_g: Species = Species.create_gas(
-        "H2O_g", solubility=solubility_models["H2O_peridotite_sossi23"]
-    )
-    H2_g: Species = Species.create_gas("H2_g")
-    O2_g: Species = Species.create_gas("O2_g")
-
-    species: SpeciesCollection = SpeciesCollection((H2O_g, H2_g, O2_g))
-    planet: Planet = Planet()
-    interior_atmosphere: InteriorAtmosphere = InteriorAtmosphere(species)
-
-    # Set up a range of fO2 shifts
-    num: int = 4
-    fO2_shifts: NumpyArrayFloat = np.linspace(-10, 10, num, dtype=np.float64)
-    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {
-        "O2_g": IronWustiteBuffer(fO2_shifts)
-    }
-
-    oceans: float = 1
-    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
-    mass_constraints: dict[str, ArrayLike] = {"H": h_kg}
-
-    interior_atmosphere.solve(
-        planet=planet,
-        fugacity_constraints=fugacity_constraints,
-        mass_constraints=mass_constraints,
-    )
-    output: Output = interior_atmosphere.output
-    solution: dict[str, ArrayLike] = output.quick_look()
-
-    target: dict[str, ArrayLike] = {
-        "H2O_g": np.array(
-            [
-                2.974916728388850e-04,
-                1.609262626310640e-01,
-                2.583020136413676e-01,
-                2.585402668639946e-01,
-            ]
-        ),
-        "H2_g": np.array(
-            [
-                2.883163373152490e01,
-                7.239157580397396e00,
-                5.393312233908341e-03,
-                2.505662833497315e-06,
-            ]
-        ),
-        "O2_g": np.array(
-            [
-                8.838513516896005e-18,
-                4.102474564576031e-11,
-                1.904200012911665e-04,
-                8.838513516896137e02,
-            ]
-        ),
-    }
-
-    assert helper.isclose(solution, target, rtol=RTOL, atol=ATOL)
-
-
-def test_H_fO2_batch_H_mass(helper) -> None:
-    """Tests H2-H2O at the IW buffer with H2O solubility for a range of H budgets."""
-
-    H2O_g: Species = Species.create_gas(
-        "H2O_g", solubility=solubility_models["H2O_peridotite_sossi23"]
-    )
-    H2_g: Species = Species.create_gas("H2_g")
-    O2_g: Species = Species.create_gas("O2_g")
-
-    species: SpeciesCollection = SpeciesCollection((H2O_g, H2_g, O2_g))
-    planet: Planet = Planet()
-    interior_atmosphere: InteriorAtmosphere = InteriorAtmosphere(species)
-
-    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {"O2_g": IronWustiteBuffer()}
-
-    oceans: float = 1
-    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
-    # Set up a range of H masses
-    mass_constraints: dict[str, ArrayLike] = {"H": np.array([h_kg, 10 * h_kg, 100 * h_kg])}
-
-    interior_atmosphere.solve(
-        planet=planet,
-        fugacity_constraints=fugacity_constraints,
-        mass_constraints=mass_constraints,
-    )
-    output: Output = interior_atmosphere.output
-    solution: dict[str, ArrayLike] = output.quick_look()
-
-    target: dict[str, ArrayLike] = {
-        "H2O_g": np.array([2.570800742364757e-01, 2.426110356931991e01, 1.610286613431932e03]),
-        "H2_g": np.array([2.491511264610584e-01, 2.351283467393216e01, 1.560621626756960e03]),
-        "O2_g": np.array([8.838513516896038e-08, 8.838513516896038e-08, 8.838513516896102e-08]),
-    }
-
-    assert helper.isclose(solution, target, rtol=RTOL, atol=ATOL)
+species: SpeciesCollection = SpeciesCollection.create(
+    ("H2_g", "H2O_g", "CO_g", "CO2_g", "CH4_g", "O2_g")
+)
+gas_CHO_system: InteriorAtmosphere = InteriorAtmosphere(species)
 
 
 def test_H_and_C(helper) -> None:
@@ -348,3 +92,213 @@ def test_H_and_C(helper) -> None:
     }
 
     assert helper.isclose(solution, target, rtol=RTOL, atol=ATOL)
+
+
+@pytest.mark.skip(reason="Checks result against previous work but not different functionality")
+def test_CHO_reduced(helper) -> None:
+    """Tests C-H-O system at IW-2
+
+    Similar to :cite:p:`BHS22{Table E, row 1}`.
+    """
+
+    planet: Planet = Planet(surface_temperature=1400)
+    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {"O2_g": IronWustiteBuffer(-2)}
+    oceans: ArrayLike = 3
+    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
+    c_kg: ArrayLike = 1 * h_kg
+    mass_constraints: dict[str, ArrayLike] = {"H": h_kg, "C": c_kg}
+
+    gas_CHO_system.solve(
+        planet=planet,
+        fugacity_constraints=fugacity_constraints,
+        mass_constraints=mass_constraints,
+    )
+    output: Output = gas_CHO_system.output
+    solution: dict[str, ArrayLike] = output.quick_look()
+
+    factsage_result: dict[str, float] = {
+        "H2_g": 175.5,
+        "H2O_g": 13.8,
+        "CO_g": 6.21,
+        "CO2_g": 0.228,
+        "CH4_g": 38.07,
+        "O2_g": 1.25e-15,
+    }
+
+    assert helper.isclose(solution, factsage_result, log=True, rtol=TOLERANCE, atol=TOLERANCE)
+
+
+def test_CHO_IW(helper) -> None:
+    """Tests C-H-O system at IW+0.5
+
+    Similar to :cite:p:`BHS22{Table E, row 2}`.
+    """
+
+    planet: Planet = Planet(surface_temperature=1400)
+    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {"O2_g": IronWustiteBuffer(0.5)}
+    oceans: ArrayLike = 3
+    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
+    c_kg: ArrayLike = 1 * h_kg
+    mass_constraints: dict[str, ArrayLike] = {"H": h_kg, "C": c_kg}
+
+    gas_CHO_system.solve(
+        planet=planet,
+        fugacity_constraints=fugacity_constraints,
+        mass_constraints=mass_constraints,
+    )
+    output: Output = gas_CHO_system.output
+    solution: dict[str, ArrayLike] = output.quick_look()
+
+    factsage_result: dict[str, float] = {
+        "CH4_g": 28.66,
+        "CO2_g": 30.88,
+        "CO_g": 46.42,
+        "H2O_g": 337.16,
+        "H2_g": 236.98,
+        "O2_g": 4.11e-13,
+    }
+
+    fastchem_result: dict[str, float] = {
+        "CH4_g": 29.61919788,
+        "CO2_g": 29.82548282,
+        "CO_g": 45.94958264,
+        "H2O_g": 332.03616807,
+        "H2_g": 236.73845646,
+        "O2_g": 3.96475584e-13,
+    }
+
+    assert helper.isclose(solution, factsage_result, log=True, rtol=TOLERANCE, atol=TOLERANCE)
+    assert helper.isclose(solution, fastchem_result, log=True, rtol=TOLERANCE, atol=TOLERANCE)
+
+
+@pytest.mark.skip(reason="Checks result against previous work but not different functionality")
+def test_CHO_oxidised(helper) -> None:
+    """Tests C-H-O system at IW+2
+
+    Similar to :cite:p:`BHS22{Table E, row 3}`.
+    """
+
+    planet: Planet = Planet(surface_temperature=1400)
+    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {"O2_g": IronWustiteBuffer(2)}
+    oceans: ArrayLike = 1
+    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
+    c_kg: ArrayLike = 0.1 * h_kg
+    mass_constraints: dict[str, ArrayLike] = {"H": h_kg, "C": c_kg}
+
+    gas_CHO_system.solve(
+        planet=planet,
+        fugacity_constraints=fugacity_constraints,
+        mass_constraints=mass_constraints,
+    )
+    output: Output = gas_CHO_system.output
+    solution: dict[str, ArrayLike] = output.quick_look()
+
+    factsage_result: dict[str, float] = {
+        "CH4_g": 0.00129,
+        "CO2_g": 3.25,
+        "CO_g": 0.873,
+        "H2O_g": 218.48,
+        "H2_g": 27.40,
+        "O2_g": 1.29e-11,
+    }
+
+    assert helper.isclose(solution, factsage_result, log=True, rtol=TOLERANCE, atol=TOLERANCE)
+
+
+@pytest.mark.skip(reason="Checks result against previous work but not different functionality")
+def test_CHO_highly_oxidised(helper) -> None:
+    """Tests C-H-O system at IW+4
+
+    Similar to :cite:p:`BHS22{Table E, row 4}`.
+    """
+
+    planet: Planet = Planet(surface_temperature=1400)
+    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {"O2_g": IronWustiteBuffer(4)}
+    oceans: ArrayLike = 1
+    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
+    c_kg: ArrayLike = 5 * h_kg
+    # Mass of O that gives the same solution as applying the buffer at IW+4
+    # o_kg: ArrayLike = 3.25196e21
+    mass_constraints: dict[str, ArrayLike] = {"H": h_kg, "C": c_kg}
+
+    gas_CHO_system.solve(
+        planet=planet,
+        fugacity_constraints=fugacity_constraints,
+        mass_constraints=mass_constraints,
+    )
+    output: Output = gas_CHO_system.output
+    solution: dict[str, ArrayLike] = output.quick_look()
+
+    factsage_result: dict[str, float] = {
+        "CH4_g": 7.13e-05,
+        "CO2_g": 357.23,
+        "CO_g": 10.21,
+        "H2O_g": 432.08,
+        "H2_g": 5.78,
+        "O2_g": 1.14e-09,
+    }
+
+    assert helper.isclose(solution, factsage_result, log=True, rtol=TOLERANCE, atol=TOLERANCE)
+
+
+def test_CHO_middle_temperature(helper) -> None:
+    """Tests C-H-O system at 873 K"""
+
+    planet: Planet = Planet(surface_temperature=873)
+    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {"O2_g": IronWustiteBuffer()}
+    oceans: ArrayLike = 1
+    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
+    c_kg: ArrayLike = 1 * h_kg
+    mass_constraints: dict[str, ArrayLike] = {"C": c_kg, "H": h_kg}
+
+    gas_CHO_system.solve(
+        planet=planet,
+        fugacity_constraints=fugacity_constraints,
+        mass_constraints=mass_constraints,
+    )
+    output: Output = gas_CHO_system.output
+    solution: dict[str, ArrayLike] = output.quick_look()
+
+    factsage_result: dict[str, float] = {
+        "H2_g": 59.066,
+        "H2O_g": 18.320,
+        "CO_g": 8.91e-4,
+        "CO2_g": 7.48e-4,
+        "CH4_g": 19.548,
+        "O2_g": 1.27e-25,
+    }
+
+    assert helper.isclose(solution, factsage_result, log=True, rtol=TOLERANCE, atol=TOLERANCE)
+
+
+def test_CHO_low_temperature(helper) -> None:
+    """Tests C-H-O system at 450 K"""
+
+    planet: Planet = Planet(surface_temperature=450)
+    fugacity_constraints: dict[str, FugacityConstraintProtocol] = {
+        "O2_g": IronWustiteBuffer(np.nan)
+    }
+    oceans: ArrayLike = 1
+    h_kg: ArrayLike = earth_oceans_to_hydrogen_mass(oceans)
+    c_kg: ArrayLike = 1 * h_kg
+    o_kg: ArrayLike = 1.02999e20
+    mass_constraints: dict[str, ArrayLike] = {"C": c_kg, "H": h_kg, "O": o_kg}
+
+    gas_CHO_system.solve(
+        planet=planet,
+        fugacity_constraints=fugacity_constraints,
+        mass_constraints=mass_constraints,
+    )
+    output: Output = gas_CHO_system.output
+    solution: dict[str, ArrayLike] = output.quick_look()
+
+    factsage_result: dict[str, float] = {
+        "H2_g": 55.475,
+        "H2O_g": 8.0,
+        "CO2_g": 1.24e-14,
+        "O2_g": 7.85e-54,
+        "CH4_g": 16.037,
+        "CO_g": 2.12e-16,
+    }
+
+    assert helper.isclose(solution, factsage_result, log=True, rtol=TOLERANCE, atol=TOLERANCE)
